Add pause feature, separate eating from moving, make preferential movement to food optional
The ability to pause between animations is useful for studying the grid
closely. That the agent has a preference to move to a food tile in its
neighbourhood is now option using "moveToFood" since we want to study
its effect. Removed setting the food tile to zero in the MoveAgents
since the eating mechanic should be separate from the moving for the
same reason.