Update MoveAgents
Updated the MoveAgents so that it loops over all alive agents and
ignores the dead ones. When the birth/death mechanic is implemented
there might be "holes" in the agentProperties matrix where an agent has
died and it hasn't been filled by a birth. Now, MoveAgents will ignore
these holes.