Add multiple experiments, moving average and progress text
Added the possibility of running multiple experiments (for statistical
purposes). At the end of the experiments, the moving average for agent
density is plotted for all experiments (in order to see long term
behaviour). The command window will now display the iteration progress
(to see how long the simulation will take)